refactor(#113): decompose console.ts into power.ts + reinstall.ts
- Extract POST /power/:poweraction, /power/restart, /power/queue/cancel → power.ts
- Extract POST /reinstall → reinstall.ts
- console.ts: 1057→438 lines, cleaner imports
